Organize proposal details to support better loading
