+++ src/components/Proposals/ProposalDetails/DisputableActions.js
@@ -9,8 +9,8 @@ import {
   VOTE_STATUS_EXECUTED,
   VOTE_STATUS_PAUSED,
   VOTE_STATUS_REJECTED,
-} from './disputable-vote-statuses'
-import { addressesEqual } from '../../lib/web3-utils'
+} from '../disputable-vote-statuses'
+import { addressesEqual } from '../../../lib/web3-utils'
 
 function DisputableActions({ status, submitter }) {
   // TODO: get connected account
