+++ src/components/Proposals/ProposalDetails/SummaryBar.js
