+++ src/components/Proposals/ProposalDetails/SummaryRow.js
