+++ src/components/Proposals/ProposalLoading.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function ProposalLoading() {
+  return <div>...Loading</div>
+}
+
+export default ProposalLoading
